# backend/openapi.json
{
  "components": {
    "schemas": {
      "ActivityEvent": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "bid_id": {
            "title": "Bid Id",
            "type": "string"
          },
          "check_name": {
            "title": "Check Name",
            "type": "string"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bid_id",
          "timestamp",
          "check_name",
          "action",
          "rule",
          "evidence"
        ],
        "title": "ActivityEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RFIDraftRequest": {
        "properties": {
          "bid_id": {
            "title": "Bid Id",
            "type": "string"
          }
        },
        "required": [
          "bid_id"
        ],
        "title": "RFIDraftRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Deterministic procurement evidence and scenario modeling.",
    "title": "PO-LICE API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/activity": {
      "get": {
        "operationId": "list_activity_api_v1_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "bid_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bid Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityEvent"
                  },
                  "title": "Response List Activity Api V1 Activity Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Activity",
        "tags": [
          "activity"
        ]
      }
    },
    "/api/v1/agent/rfi-draft": {
      "post": {
        "description": "Generate and persist a deterministic RFI draft from patrol findings.\n\nThe draft is always created with status=DRAFT and human_reviewed=False.\nApproval is a separate action via PATCH /api/v1/rfis/{rfi_id}/approve.",
        "operationId": "generate_rfi_draft_api_v1_agent_rfi_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFIDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFIDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Rfi Draft",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/v1/bids/simulate": {
      "post": {
        "description": "Recalculate the Dynamic Docket from bounded, deterministic inputs.",
        "operationId": "simulate_bid_api_v1_bids_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Bid",
        "tags": [
          "bids"
        ]
      }
    },
    "/api/v1/bids/upload": {
      "post": {
        "description": "Validate and analyse one PDF. Source text is extracted; patrols make the decision.",
        "operationId": "upload_and_audit_bid_api_v1_bids_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_and_audit_bid_api_v1_bids_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload And Audit Bid",
        "tags": [
          "bids"
        ]
      }
    },
    "/api/v1/bids/{bid_id}": {
      "delete": {
        "operationId": "delete_bid_api_v1_bids__bid_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "title": "Bid Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bid",
        "tags": [
          "bids"
        ]
      },
      "get": {
        "operationId": "get_bid_api_v1_bids__bid_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bid",
        "tags": [
          "bids"
        ]
      }
    },
    "/api/v1/bids/{bid_id}/actions": {
      "post": {
        "operationId": "record_reviewer_action_api_v1_bids__bid_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewerActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Reviewer Action",
        "tags": [
          "bids"
        ]
      }
    },
...
    "/api/v1/bids/{bid_id}/source": {
      "get": {
        "operationId": "download_source_api_v1_bids__bid_id__source_get",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Source",
        "tags": [
          "bids"
        ]
      }
    },
...
  }
}